+++ src/ebus_mqtt_client/__init__.py
@@ -1,5 +1,5 @@
 from ebus_mqtt_client.client import AUTH_TYPE_USER_PASS, MqttClient
 
-__version__ = "0.1.7"
+__version__ = "0.1.8"
 
 __all__ = ["MqttClient", "AUTH_TYPE_USER_PASS"]